requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
llama-index-core==0.11.20
llama-index-vector-stores-postgres==0.2.5
llama-index-embeddings-huggingface==0.3.0
sentence-transformers==3.1.1
litellm==1.51.3
psycopg==3.2.3
celery==5.4.0
redis==5.2.1
pydantic==2.9.2
python-multipart==0.0.9
pymupdf==1.24.10
pandas==2.2.3
sqlalchemy==2.0.35